requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
python-multipart>=0.0.9,<1.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=4.0,<5.0
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
email-validator>=2.1,<3.0
openai>=1.35,<2.0
google-generativeai>=0.7,<1.0
faiss-cpu>=1.8,<2.0
numpy>=1.26,<3.0
PyMuPDF>=1.24,<2.0
pytesseract>=0.3,<1.0
Pillow>=10.3,<12.0
